Reorganize package structure for velocity module
