+++ velocity/src/main/java/net/thenextlvl/commander/command/CommanderCommand.java
@@ -1,10 +1,10 @@
-package net.thenextlvl.commander.velocity.command;
+package net.thenextlvl.commander.command;
 
 import com.velocitypowered.api.command.CommandSource;
 import com.velocitypowered.api.command.SimpleCommand;
 import lombok.RequiredArgsConstructor;
 import net.kyori.adventure.text.minimessage.tag.resolver.Placeholder;
-import net.thenextlvl.commander.velocity.CommanderPlugin;
+import net.thenextlvl.commander.CommanderPlugin;
 
 import java.util.ArrayList;
 import java.util.Collections;
