+++ velocity/src/main/java/net/thenextlvl/commander/listener/package-info.java
@@ -2,7 +2,7 @@
 @FieldsAreNotNullByDefault
 @MethodsReturnNotNullByDefault
 @ParametersAreNotNullByDefault
-package net.thenextlvl.commander.velocity.implementation;
+package net.thenextlvl.commander.listener;
 
 import core.annotation.FieldsAreNotNullByDefault;
 import core.annotation.MethodsReturnNotNullByDefault;
